+++ integrations/attio/definitions/actions.ts
@@ -1,11 +1,5 @@
-import { z, type ActionDefinition } from '@botpress/sdk'
-
-export const baseIdentifierSchema = z
-  .object({
-    workspace_id: z.string().title('Workspace ID').describe('The Attio workspace ID'),
-    object_id: z.string().title('Object ID').describe('The Attio object ID'),
-  })
-  .title('Record Identifier')
+import { ActionDefinition, z } from '@botpress/sdk'
+import { baseIdentifierSchema } from './common'
 
 const recordSchema = z
   .object({
